refactor: create package errorutil
- implement errorutil.WithDetails
- replace local implementations
